Add files via upload
In the Portfolio class, the program begins by presenting a menu that showcases various data structures and algorithms. Users are then prompted to select an option from this menu. Depending on the user's choice, the program either invokes the main method of other classes (such as JavaList, LinkedList, stackOperations, cipherText, Trees, BinaryTree, Graph) or gracefully exits. This cycle continues until the user decides to terminate the program. The application employs a menu-based strategy for engaging with diverse data structures and algorithms. Some sections feature input validation to manage user inputs that may be invalid. The program offers an organized method for exploring and interacting with a range of data structures and algorithms. Each class is assigned specific responsibilities, contributing to a modular and well-organized code structure.